Propagate monster names via CVars
Because server's customized names doesn't get sync'ed to clients unless
it is defined in CVARINFO, and ACS isn't able to send dynamic strings
over the network. We simply store the name in predefined server CVars
which gets sync'ed by Zandronum itself. It is a terrible workaround, but
then there probably isn't any way around this.